Add the new options component
