+++ src/models/MainWindowModel.cs
@@ -4,27 +4,18 @@ namespace Filters.Models
 {
     public class MainWindowModel : ReactiveObject
     {
-        // Path for the image
-        private string path;
-        public string Path
+        // Options for the output
+        private OptionsModel options;
+        public OptionsModel Options
         {
-            get => path;
-            set => this.RaiseAndSetIfChanged(ref path, value);
-        }
-
-        // True for manual input for the path
-        private bool isManual;
-        public bool IsManual
-        {
-            get => isManual;
-            set => this.RaiseAndSetIfChanged(ref isManual, value);
+            get => options;
+            set => this.RaiseAndSetIfChanged(ref options, value);
         }
 
         /// <summary>Constructor</summary>
         public MainWindowModel()
         {
-            path = "No image. Please select one.";
-            isManual = false;
+            options = new OptionsModel();
         }
     }
 }
