+++ src/views/main/MainWindow.xaml
@@ -2,6 +2,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:options="clr-namespace:Filters.Views.Options"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
         x:Class="Filters.Views.Main.MainWindow"
         Width="980"
@@ -43,43 +44,7 @@
             <Border Classes="Card"
                     Padding="3">
                 <StackPanel>
-                    <TextBlock
-                               HorizontalAlignment="Center"
-                               Margin="5"
-                               Text="Select the way to chose the image." />
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="*" />
-                            <ColumnDefinition Width="*" />
-                        </Grid.ColumnDefinitions>
-                        <RadioButton Click="SelectChooseMethod"
-                                     Content="File Chooser"
-                                     Grid.Column="0"
-                                     HorizontalAlignment="Center"
-                                     IsChecked="True"
-                                     Name="Chooser"/>
-                        <RadioButton Click="SelectChooseMethod"
-                                     Content="Manual Write"
-                                     Grid.Column="1"
-                                     HorizontalAlignment="Center"
-                                     Name="Manual"/>
-                    </Grid>
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="0.15*" />
-                            <ColumnDefinition Width="0.85*" />
-                        </Grid.ColumnDefinitions>
-                        <Button Click="ChooseImage"
-                                Content="Choose Image"
-                                Grid.Column="0"
-                                IsVisible="{Binding !IsManual}"
-                                Margin="10" />
-                        <TextBox Grid.Column="1"
-                                 IsReadOnly="{Binding !IsManual}"
-                                 Margin="0,0,75,0"
-                                 VerticalAlignment="Center"
-                                 Text="{Binding Path}" />
-                    </Grid>
+                    <options:Options />
                     <Grid>
                         <Grid.ColumnDefinitions>
                             <ColumnDefinition Width="0.50*" />
