+++ src/views/main/MainWindow.xaml.cs
@@ -14,51 +14,10 @@ namespace Filters.Views.Main
             InitializeComponent();
         }
 
-        /// <summary>Let the user choose an image and put the path on screen</summary>
-        /// <param name="sender">The object that raised the event</param>
-        /// <param name="e">The object that is being handled</param>
-        private async void ChooseImage(object sender, RoutedEventArgs e)
-        {
-            string path = await GetPath();
-            MainWindowModel context = DataContext as MainWindowModel;
-            context.Path = path;
-        }
-
-        /// <summary>Open the file chooser</summary>
-        /// <returns>The path of the image selected</returns>
-        private async Task<string> GetPath()
-        {
-            OpenFileDialog dialog = new OpenFileDialog
-            {
-                Title = "Choose Image",
-                AllowMultiple = false
-            };
-            dialog.Filters.Add(new FileDialogFilter { Name = "Image", Extensions = { "png" } });
-
-            string[] result = await dialog.ShowAsync(this);
-            if (result.Length > 0)
-                return result[0];
-            return "No image. Please select one.";
-        }
-
         /// <summary>Initialize all the components</summary>
         private void InitializeComponent()
         {
             AvaloniaXamlLoader.Load(this);
-        }
-
-        /// <summary>Display the controls based on the option selected</summary>
-        /// <param name="sender">The object that raised the event</param>
-        /// <param name="e">The object that is being handled</param>
-        private void SelectChooseMethod(object sender, RoutedEventArgs e)
-        {
-            RadioButton radioButton = sender as RadioButton;
-            MainWindowModel context = DataContext as MainWindowModel;
-
-            if (radioButton.Name.Equals("Manual"))
-                context.IsManual = true;
-            else
-                context.IsManual = false;
         }
     }
 }
